MAINT: rename files for consistency
