+++ q2_boots/__init__.py
@@ -11,9 +11,9 @@ from ._alpha import (alpha,
                      alpha_collection,
                      alpha_average)
 from ._version import get_versions
-from .beta import (beta, get_medoid, beta_collection, per_cell_average,
-                   beta_average)
-from .core_metrics import (core_metrics)
+from ._beta import (beta, get_medoid, beta_collection, per_cell_average,
+                    beta_average)
+from ._core_metrics import (core_metrics)
 from . import _version
 
 __all__ = ['resample',
